pyproject.toml
[project]
name = "home-credit-default-risk"
version = "0.1.0"
description = "Home Credit Default Risk — target Private LB ROC AUC ≥ 0.78 (stretch ≥ 0.80)"
authors = [{ name = "Aly Ayman" }]
requires-python = ">=3.11,<3.12"
readme = "README.md"
license = { text = "MIT" }

dependencies = [
    # Data
    "polars>=1.12.0",
    "pandas>=2.2.0",
    "pyarrow>=18.0.0",
    "numpy>=1.26.0,<2.0.0",
    # ML / GBM
    "lightgbm>=4.5.0",
    "xgboost>=2.1.0",
    "catboost>=1.2.7",
    "scikit-learn>=1.5.0",
    "scipy>=1.13.0",
    # NN
    "torch>=2.4.0",
    # Tuning
    "optuna>=4.0.0",
    # Utils
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
    "joblib>=1.4.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.7.0",
    "black>=24.10.0",
    "isort>=5.13.0",
    "pytest>=8.3.0",
    "pytest-cov>=5.0.0",
    "ipykernel>=6.29.0",
    "jupyter>=1.1.0",
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
]

[tool.uv]
# Use uv.lock for reproducibility. CUDA torch wheels resolved per-platform via the official PyTorch index.
# To install GPU torch on Windows/Linux:
#   uv pip install torch --index-url https://download.pytorch.org/whl/cu121
# Then pin with `uv lock`.

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

# ─── Linting / Formatting ─────────────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py311"
exclude = ["notebooks/", "data/", "artifacts/", "submissions/"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM", "RET"]
ignore = ["E501", "B008"]

[tool.black]
line-length = 100
target-version = ["py311"]
exclude = '/(notebooks|data|artifacts|submissions)/'

[tool.isort]
profile = "black"
line_length = 100
skip = ["notebooks", "data", "artifacts", "submissions"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
